requirements.txt
qrcode>=7.3.1
pillow>=9.0.0
opencv-python>=4.5.5
pyzbar>=0.1.9
numpy>=1.22.0 